--- lib/src/flips.dart
@@ -1,28 +1,0 @@
-part of tiled;
-
-class Flips {
-  final bool horizontally;
-  final bool vertically;
-  final bool diagonally;
-
-  const Flips(this.horizontally, this.vertically, this.diagonally);
-
-  const Flips.defaults() : this(false, false, false);
-
-  Flips copyWith({
-    bool horizontally,
-    bool vertically,
-    bool diagonally,
-  }) {
-    return Flips(
-      horizontally ?? this.horizontally,
-      vertically ?? this.vertically,
-      diagonally ?? this.diagonally,
-    );
-  }
-
-  @override
-  String toString() {
-    return 'Flips { horizontally: $horizontally, vertically: $vertically, diagonally: $diagonally }';
-  }
-}
