--- lib/src/json/propertyjson.dart
@@ -1,29 +1,0 @@
-import 'package:xml/xml.dart';
-
-class PropertyJson {
-  String name;
-  String type;
-  String value;
-
-  PropertyJson({this.name, this.type, this.value});
-
-  PropertyJson.fromXML(XmlNode element) {
-    value = element.getAttribute('value');
-    name = element.getAttribute('name');
-    type = element.getAttribute('type');
-  }
-
-  PropertyJson.fromJson(Map<String, dynamic> json) {
-    name = json['name'];
-    type = json['type'];
-    value = json['value'].toString();
-  }
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = <String, dynamic>{};
-    data['name'] = name;
-    data['type'] = type;
-    data['value'] = value;
-    return data;
-  }
-}
