--- lib/src/tsx_provider.dart
@@ -1,5 +1,0 @@
-part of tiled;
-
-abstract class TsxProvider {
-  String getSource(String key);
-}
