+++ lib/tiled.dart
@@ -1,25 +1,29 @@
 library tiled;
 
-import 'dart:io';
-import 'dart:math';
 import 'dart:convert';
+import 'dart:io';
 import 'dart:typed_data';
 
-import 'package:tiled/src/json/mapjson.dart';
-import 'package:xml/xml.dart';
 import 'package:archive/archive.dart';
+import 'package:xml/xml.dart';
 
-part 'src/tile_map_parser.dart';
-part 'src/json/tile_map_json_parser.dart';
-part 'src/tile_map.dart';
-part 'src/tileset.dart';
-part 'src/image.dart';
+part 'src/chunk.dart';
+part 'src/editorsetting.dart';
+part 'src/export.dart';
+part 'src/frame.dart';
+part 'src/grid.dart';
 part 'src/layer.dart';
+part 'src/tiledobject.dart';
+part 'src/point.dart';
+part 'src/property.dart';
+part 'src/terrain.dart';
+part 'src/text.dart';
 part 'src/tile.dart';
-part 'src/object_group.dart';
-part 'src/tmx_object.dart';
-part 'src/node_dsl.dart';
-part 'src/tsx_provider.dart';
-part 'src/flips.dart';
-
-XmlDocument _parseXml(String input) => XmlDocument.parse(input);
+part 'src/tile_map_parser.dart';
+part 'src/tiledimage.dart';
+part 'src/tiledmap.dart';
+part 'src/tileoffset.dart';
+part 'src/tileset.dart';
+part 'src/wangcolor.dart';
+part 'src/wangset.dart';
+part 'src/wangtile.dart';
