+++ test/object_group_test.dart
@@ -1,5 +1,4 @@
 import 'package:test/test.dart';
-import 'package:tiled/tiled.dart';
 import 'package:xml/xml.dart';
 import 'dart:io';
