+++ test/parser_compare_test.dart
@@ -1,50 +1,57 @@
 import 'dart:io';
 
 import 'package:test/test.dart';
-import 'package:tiled/src/json/mapjson.dart';
 import 'package:tiled/tiled.dart';
 
 void main() {
-  final parser = TileMapJsonParser();
 
-  MapJson map;
+  TiledMap map;
   setUp(() {
     return File('./test/fixtures/testcsv.json').readAsString().then((xml) {
-      map = parser.parse(xml);
+      map = TileMapParser.parseJson(xml);
       // var out = new File('jsonoutcome.json').openWrite();
       // out.write(map.toTileMap());
       // out.close();
     });
   });
 
-  MapJson map2;
+  TiledMap map2;
   setUp(() {
     return File('./test/fixtures/testzlib.json').readAsString().then((xml) {
-      map2 = parser.parse(xml);
+      map2 = TileMapParser.parseJson(xml);
     });
   });
 
 
-  MapJson map3;
+  TiledMap map3;
   setUp(() {
     return File('./test/fixtures/testgzip.json').readAsString().then((xml) {
-      map3 = parser.parse(xml);
+      map3 = TileMapParser.parseJson(xml);
     });
   });
 
 
-  MapJson map4;
+  TiledMap map4;
   setUp(() {
     return File('./test/fixtures/testbase64only.json').readAsString().then((xml) {
-      map4 = parser.parse(xml);
+      map4 = TileMapParser.parseJson(xml);
     });
   });
 
-  final tileparser = TileMapParser();
-  TileMap tilemap;
+  TiledMap tilemap;
   setUp(() {
     return File('./test/fixtures/test.tmx').readAsString().then((xml) {
-      tilemap = tileparser.parse(xml);
+      tilemap = TileMapParser.parseTmx(xml);
+      // var out = new File('tmxoutcome.json').openWrite();
+      // out.write(tilemap);
+      // out.close();
+    });
+  });
+
+  TiledMap tilemapEllipse;
+  setUp(() {
+    return File('./test/fixtures/map.tmx').readAsString().then((xml) {
+      tilemapEllipse = TileMapParser.parseTmx(xml);
       // var out = new File('tmxoutcome.json').openWrite();
       // out.write(tilemap);
       // out.close();
@@ -52,36 +59,36 @@ void main() {
   });
 
 
-  MapJson map5;
+  TiledMap map5;
   setUp(() {
     return File('./test/fixtures/complexmap.json').readAsString().then((xml) {
-      map5 = parser.parse(xml);
+      map5 = TileMapParser.parseJson(xml);
     });
   });
 
-  MapJson map6;
+  TiledMap map6;
   setUp(() {
     return File('./test/fixtures/complexmap.tmx').readAsString().then((xml) {
-      map6 = parser.parseXml(xml);
+      map6 = TileMapParser.parseTmx(xml);
     });
   });
 
-  final tileparserComplex = TileMapParser();
-  TileMap tilemapComplex;
+  TiledMap tilemapComplex;
   setUp(() {
     return File('./test/fixtures/complexmap.tmx').readAsString().then((xml) {
-      tilemapComplex = tileparserComplex.parse(xml);
+      tilemapComplex = TileMapParser.parseTmx(xml);
     });
   });
 
   group('Parser compare', () {
-    test('toString should be equal', () => expect(map.toTileMap().toString(), equals(tilemap.toString())));
-    test('toString should be equal', () => expect(map2.toTileMap().toString(), equals(tilemap.toString())));
-    test('toString should be equal', () => expect(map3.toTileMap().toString(), equals(tilemap.toString())));
-    test('toString should be equal', () => expect(map4.toTileMap().toString(), equals(tilemap.toString())));
+    test('toString should be equal', () => expect(map, equals(tilemap)));
+    test('toString should be equal', () => expect(map2, equals(tilemap)));
+    test('toString should be equal', () => expect(map3, equals(tilemap)));
+    test('toString should be equal', () => expect(map4, equals(tilemap)));
 
-    test('toString should be equal', () => expect(map5.toTileMap().toString(), equals(tilemapComplex.toString())));
-    test('toString should be equal', () => expect(map6.toTileMap().toString(), equals(tilemapComplex.toString())));
+    test('toString should be equal', () => expect(map5, equals(tilemapComplex)));
+    test('toString should be equal', () => expect(map6, equals(tilemapComplex)));
+    test('toString should be equal', () => expect(tilemapEllipse, equals(tilemapComplex)));
   });
 
 }
